pyproject.toml
[tool.poetry]
name = "hackathon-gaia"
version = "0.1.0"
description = "Hackathon sur l'IA générative appliquée à l'agriculture - février 2024"
authors = ["perrine.lepelley <perrine.lepelley@eureden.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.0"
langchain = "^0.1.7"
langchain-core = "^0.1.23"
langchain-mistralai = "^0.0.4"
langchain-community = "^0.0.20"
faiss-cpu = "^1.7.4"
ipykernel = "^6.29.2"
sentence-transformers = "^2.3.1"
python-dotenv = "^1.0.1"
chainlit = "^1.0.0"
pypdf = "^4.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
